+++ enterprise/backend/src/baserow_enterprise/auth_provider/exceptions.py
@@ -1,0 +1,5 @@
+class DifferentAuthProvider(Exception):
+    """
+    Raised when logging in an existing user that should not
+    be logged in using a different than the approved auth provider.
+    """